整体调整。加入画光插件。

+++ src/libs/common/vec.h
@@ -10,13 +10,13 @@
 #include<math.h>
 class Vec3{
 public:
-    float x,y,z;
+    double x,y,z;
 
     //! 默认构造函数,不执行任何操作
     Vec3(){}
     //! 复制构造函数
     Vec3(const Vec3 &a):x(a.x),y(a.y),z(a.z){}
-    Vec3(float nx,float ny,float nz):x(nx),y(ny),z(nz){}
+    Vec3(double nx,double ny,double nz):x(nx),y(ny),z(nz){}
 
     ///标准对象操作
     //! 重载赋值运算符，返回引用实现左值。
@@ -45,11 +45,11 @@ public:
         return Vec3(x-a.x,y-a.y,z-a.z);
     }
     //与标量的乘除法
-    Vec3 operator * (float a)const{
+    Vec3 operator * (double a)const{
         return Vec3(x*a,y*a,z*a);
     }
-    Vec3 operator / (float a)const{
-        float oneOverA =1.0f/a; //不对除0进行检查
+    Vec3 operator / (double a)const{
+        double oneOverA =1.0f/a; //不对除0进行检查
         return Vec3(x*oneOverA,y*oneOverA,z*oneOverA);
     }
     //重载自反运算符
@@ -61,12 +61,12 @@ public:
         x -=a.x; y-=a.y; z-=a.z;
         return *this;
     }
-    Vec3 &operator *=(float a){
+    Vec3 &operator *=(double a){
         x*=a;y*=a;z*=a;
         return *this;
     }
-    Vec3 &operator /=(float a){
-        float oneOverA =1.0f/a; //不对除0进行检查
+    Vec3 &operator /=(double a){
+        double oneOverA =1.0f/a; //不对除0进行检查
         x*=oneOverA; y*=oneOverA; z*=oneOverA;
         return *this;
     }
@@ -77,22 +77,22 @@ public:
     }
     //向量标准化
     void normalize(){
-        float magSq=x*x+y*y+z*z;
+        double magSq=x*x+y*y+z*z;
         if(magSq>0.0f){//检查除0
-            float oneOverMag=1.0f/sqrt(magSq);
+            double oneOverMag=1.0f/sqrt(magSq);
             x*=oneOverMag;
             y*=oneOverMag;
             z*=oneOverMag;
         }
     }
     //向量点乘，重载标准乘法运算符
-    float operator *(const Vec3 &a)const{
+    double operator *(const Vec3 &a)const{
         return x*a.x+y*a.y+z*a.z;
     }
 
     ///成员函数版本
     //向量的模
-    float mag(){
+    double mag(){
         return sqrt(x*x+y*y+z*z);
     }
 
@@ -123,7 +123,7 @@ public:
 
 ///vector3非成员函数
 //! 求向量模，非成员函数版本
-inline float vectorMag(const Vec3 &a){
+inline double vectorMag(const Vec3 &a){
     return sqrt(a.x*a.x+a.y*a.y+a.z*a.z);
 }
 //! 计算两向量叉乘
@@ -135,14 +135,14 @@ inline Vec3 crossProduct(const Vec3 &a,const Vec3 &b){
                 );
 }
 //! 实现标量左乘
-inline Vec3 operator *(float k, const Vec3 &v){
+inline Vec3 operator *(double k, const Vec3 &v){
     return Vec3(k*v.x,k*v.y,k*v.z);
 }
 //! 计算两点间距离
-inline float distance(const Vec3 &a,const Vec3 &b){
-    float dx=a.x-b.x;
-    float dy=a.y-b.y;
-    float dz=a.z-b.z;
+inline double distance(const Vec3 &a,const Vec3 &b){
+    double dx=a.x-b.x;
+    double dy=a.y-b.y;
+    double dz=a.z-b.z;
     return sqrt(dx*dx+dy*dy+dz*dz);
 }
